requirements.txt
python-dotenv
python-telegram-bot==20.3
transformers
torch
whisper
pydub
